+++ src/domain/body/fallingBody/Remains.java
@@ -4,10 +4,10 @@ public class Remains extends FallingBody {
 
     int remainNumber;
 
-    public Remains(double x_coordinates,
-                   double y_coordinates,
-                   double length,
-                   double width,
+    public Remains(int x_coordinates,
+                   int y_coordinates,
+                   int length,
+                   int width,
                    int remainNumber) {
         super(x_coordinates, y_coordinates, length, width);
         this.remainNumber=remainNumber;
